requirements/base.txt
Django==1.7.1
djangorestframework==3.1.0
Markdown==2.5.2
django-filter==0.9.1
django-oauth-plus==2.2.5
django-oauth2-provider==0.2.6.1
django-guardian==1.2.4
pygments==2.0.1
